# openapi.json
{
    "components": {
        "schemas": {
...
            "ErrorDetail": {
                "additionalProperties": false,
                "properties": {
                    "location": {
                        "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
                        "type": "string"
                    },
                    "message": {
                        "description": "Error message text",
                        "type": "string"
                    },
                    "value": {
                        "description": "The value at the given location"
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ErrorModel.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "detail": {
                        "description": "A human-readable explanation specific to this occurrence of the problem.",
                        "examples": [
                            "Property foo is required but is missing."
                        ],
                        "type": "string"
                    },
                    "errors": {
                        "description": "Optional list of individual error details",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    },
                    "instance": {
                        "description": "A URI reference that identifies the specific occurrence of the problem.",
                        "examples": [
                            "https://example.com/error-log/abc123"
                        ],
                        "format": "uri",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP status code",
                        "examples": [
                            400
                        ],
                        "format": "int64",
                        "type": "integer"
                    },
                    "title": {
                        "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
                        "examples": [
                            "Bad Request"
                        ],
                        "type": "string"
                    },
                    "type": {
                        "default": "about:blank",
                        "description": "A URI reference to human-readable documentation for the error.",
                        "examples": [
                            "https://example.com/errors/example"
                        ],
                        "format": "uri",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PublicShareOutBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/PublicShareOutBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "created": {
                        "description": "RFC3339 timestamp the token was issued",
                        "type": "string"
                    },
                    "token": {
                        "description": "32-character hex share token",
                        "type": "string"
                    },
                    "url": {
                        "description": "Shareable URL path (/public/<token>)",
                        "type": "string"
                    }
                },
                "required": [
                    "token",
                    "url",
                    "created"
                ],
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "weft-loom-server API",
        "version": "v1"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/projects/{name}/public-share": {
            "delete": {
                "description": "Removes the sidecar + drops the token from the in-memory index. Idempotent \u2014 returns 204 even when no share existed.",
                "operationId": "delete-public-share",
                "parameters": [
                    {
                        "description": "Project name (filesystem-safe identifier)",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "description": "Project name (filesystem-safe identifier)",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "summary": "Revoke the public-share token for the project",
                "tags": [
                    "public-share"
                ]
            },
            "get": {
                "description": "Returns the token + shareable URL when one exists ; 404 when the project has no active share.",
                "operationId": "get-public-share",
                "parameters": [
                    {
                        "description": "Project name (filesystem-safe identifier)",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "description": "Project name (filesystem-safe identifier)",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicShareOutBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "summary": "Read the current public-share token for the project",
                "tags": [
                    "public-share"
                ]
            },
            "post": {
                "description": "Generates a fresh 32-hex-char token + persists the sidecar. If a token already existed for this project it is replaced (the prior URL stops working). Returns the new token, its shareable URL, and the RFC3339 issue time.",
                "operationId": "create-public-share",
                "parameters": [
                    {
                        "description": "Project name (filesystem-safe identifier)",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "description": "Project name (filesystem-safe identifier)",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublicShareOutBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "summary": "Issue (or rotate) a public-share token for the project",
                "tags": [
                    "public-share"
                ]
            }
        },
...
    }
}